requirements.txt
playwright
pandas
gspread
google-auth
